rename: 공고 페이지와 컴포넌트 파일의 분리(#6)
